move ClipToPostNitroGenesis to blockchain & use params
